// dxfCreator/src/types.ts
export type EquipmentKind =
  | 'bess'
  | 'inverter'
  | 'aux-transformer'
  | 'aux-distribution'
  | 'fiber-junction-box'
  | 'junction-box'
  | 'substation';

export type CableClass =
  | 'bess-feeder-14a1'
  | 'bess-feeder-14a2'
  | 'bess-feeder-14b1'
  | 'bess-feeder-14b2'
  | 'bess-feeder-15a1'
  | 'bess-feeder-15a2'
  | 'bess-feeder-15b1'
  | 'bess-feeder-15b2'
  | 'aux-feeder'
  | 'dc-negative'
  | 'dc-positive'
  | 'fiber';

export type TrenchClass = 'mvac' | 'dc' | 'aux-fiber';